+++ lib/src/internals/indicator_wrap.dart
@@ -91,6 +91,10 @@ abstract class RefreshIndicatorState<T extends RefreshIndicator>
       mode = RefreshStatus.idle;
     }
 
+    // If FrontStyle overScroll,it shouldn't disable gesture in scrollable
+    if(_position.extentBefore==0.0&&widget.refreshStyle==RefreshStyle.Front) {
+      _position.context.setIgnorePointer(false);
+    }
     // Sometimes different devices return velocity differently, so it's impossible to judge from velocity whether the user
     // has invoked animateTo (0.0) or the user is dragging the view.Sometimes animateTo (0.0) does not return velocity = 0.0
     // velocity < 0.0 may be spring up,>0.0 spring down